backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic>=2.5,<3
python-dotenv>=1.0
httpx>=0.27
supabase>=2.6
